+++ src/Definition/BreadcrumbDefinition.php
@@ -10,7 +10,8 @@ final class BreadcrumbDefinition
         public readonly array $routeParameters = [],
         public readonly bool $routeAbsolute = true,
         public readonly int $position = 0,
-        public readonly array $attributes = []
+        public readonly array $attributes = [],
+        public readonly ?string $parentRoute = null
     )
     {
     }