+++ src/Loader/BreadcrumbContext.php
@@ -6,10 +6,14 @@ use Symfony\Component\HttpFoundation\Request;
 
 final class BreadcrumbContext
 {
+    public readonly ?string $routeName;
+
     public function __construct(
         public readonly Request $request,
-        public readonly mixed $controller
+        public readonly mixed $controller,
+        ?string $routeName = null
     )
     {
+        $this->routeName = $routeName ?? $request->attributes->get('_route');
     }
 }